Basic Box2D environment
Using box2d a basic world is set up
With a simple box body joint with a leg.
The leg is another two boxes joint together.

Added a contact listern which invokes game over
if body and ground comes in contact.

Requires further work with motor and joints.
Need to add a way to indentify state.